allowing switching of async between py2 and py3
Hi John,

In my init file I have something like,

```elisp
(defun switch-to-py3 ()
  (interactive)
  (setq org-babel-python-command "python3")
  (setq python-shell-interpreter "python3")
  (setq python-shell-buffer-name "Python3")
  )

(defun switch-to-py2 ()
  (interactive)
  (setq org-babel-python-command "python2")
  (setq python-shell-interpreter "python2")
  (setq python-shell-buffer-name "Python2")
  )
```

to allow me to switch between py2 and py3 between projects. This works well for the default code execution with C-c C-c, but not with async python. The change in this PR seems to fix that.