fix for issue #68
Apparently cl-flet is not working right now to fix the level problem. This fix
replaces it with flet which does work. I also added a function to add extra \n
to strings. A single \n isn't enough to start a new paragraph in the notebook,
you need two of them.