address part of issue #113
For some reason, there is not always a -> indicator on the problematic line, so
the regexp does not always work. In cases where it does not work, we just go to
the origin of the src block now.

It looks like this commit also introduces a lot of boring indentation changes.