This is significant change in ipython for scimax
The upstream ob-ipython has fixed some issues, so this branch adds it for scimax
to use instead of the customized fork it had been using.

Notably, sync/async is much simpler now.

I refactored some code out into scimax-ob.el that was not specific to ipython.
There is more to move there still from the old scimax-org-babel-ipython.el code.

Doing it this way probably will not be disruptive to people that like the old
way, but that is eventually going to be removed I think.