WIP: edit/command hydras
The idea is these have the same bindings as Jupyter. There are multiple hydras
now. I am not sure if they should all stay. On one hand, this set is almost like
the jupyter bindings, which should be good for jupyter users. They are kind of
constrained to that though, by design. On the other hand, the other one is free
to be designed and include what you want, at the risk of doing things jupyter
users are unfamiliar with. Maybe it is possible to integrate them without
causing problems.

They need a good hint bar.