Small fix for issue #167.
This uses an image path if it exists, and otherwise expands to a temp dir path.
I don't use this code often. This fixes the wrong paths used in the tempdir
expansion, but I am not sure if it breaks anything else. It is probably worth
figuring out if this is the best way to support html mail. I forked org-mime a
long time ago to fix some issues that may have since been addressed.