Resolve issue #199
This commit helps ob-ipython respect :dir header args on src blocks. It should
not change behavior if no :dir header is set.

If there is a :dir header, then this stores the cwd, changes to the :dir
directory, runs the code, and changes back to the original cwd. This allows you
set :dir on subtrees, but if you make a new heading the :dir from the previous
subtree does not impact the new heading.

This solution is adapted from one suggested by @dangom. I added the change back
to the cwd at the end.