fix issue when jupyter is slow to start
I found it frequently takes more than one second for the jupyter console process
to start, and then a json file is not available when the driver.py file needs
it, leading to a startup error. I modified the create-kernel function here to
wait longer (ideally about long enough) to avoid this. It seems to work more
often than it did before, so I am committing this. Sometimes, it seems like it
gets frozen, but that too is intermittent.