Improve flexibility on showing execution count
I changed the defcustom to be a function. We don't use
ob-ipython-suppress-execution-count anymore.

Now you use a function that returns an empty string to suppress output. There
are example functions for putting the execution count in an overlay in the
margin, in an attribute, in the output, and suppressing it.