remove +smart for docx
in issue #338 this was reported an issue. I confirmed this is not currently
supported for docx. I assume at one time it was, but in my current versions it
doesn't appear to be.

pandoc -v
pandoc 2.2.3.2
Compiled with pandoc-types 1.17.5.1, texmath 0.11.0.1, skylighting 0.7.2
Default user data directory: /Users/jkitchin/.pandoc
Copyright (C) 2006-2018 John MacFarlane
Web:  http://pandoc.org
This is free software; see the source for copying conditions.
There is no warranty, not even for merchantability or fitness
for a particular purpose.