add customization for the spell check language.
Reported in issue #348. We shouldn't hard code this choice here. Maybe an
alternative is to just pass this completely on to the user. It is mostly here
for convenience (for me). This makes it at least more easily configurable.
Although, a user can always change it in the user.el file with another setq.