remove macro for f-string and replace by function
I don't recall why I made this a macro, probably because of all the code
building in it. I just eval the last item now. I don't recall if that has some
tricky namespace/env issues that make a macro better.

I changed it because it causes an Eager expansion message for scimax-utils,
reported in issue #366.