issue #464 potential fix
It seems that imagemagick may no longer be used in emacs 28+ by default. These
lines are old, so to start, I moved them out of scimax-org and into
scimax-org-images for consolidation, and then commented them out. It seems
possible we don't need them anymore.

It seems ok to leave the inhibit line in, it should have no effect if
imagemagick is not installed.